+++ backend/controllers/ReviewController.php
@@ -23,6 +23,7 @@ class ReviewController extends Controller
         return [
             'access' => [
                 'class' => AccessControl::class,
+                'only' => ['index', 'view', 'create', 'update', 'delete'],
                 'rules' => [
                     [
                         'allow' => true,
@@ -45,6 +46,12 @@ class ReviewController extends Controller
      */
     public function actionIndex()
     {
+        if (!Yii::$app->user->can('viewReview')) {
+            throw new \yii\web\ForbiddenHttpException(
+                'You do not have permission to view reviews.'
+            );
+        }
+
         $searchModel = new ReviewSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -62,6 +69,12 @@ class ReviewController extends Controller
      */
     public function actionView($id_review)
     {
+        if (!Yii::$app->user->can('viewReview')) {
+            throw new \yii\web\ForbiddenHttpException(
+                'You do not have permission to view this review.'
+            );
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id_review),
         ]);
@@ -74,6 +87,12 @@ class ReviewController extends Controller
      */
     public function actionCreate()
     {
+        if (!Yii::$app->user->can('createReview')) {
+            throw new \yii\web\ForbiddenHttpException(
+                'You do not have permission to create reviews.'
+            );
+        }
+
         $model = new Review();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -94,6 +113,12 @@ class ReviewController extends Controller
      */
     public function actionUpdate($id_review)
     {
+        if (!Yii::$app->user->can('updateReview')) {
+            throw new \yii\web\ForbiddenHttpException(
+                'You do not have permission to update reviews.'
+            );
+        }
+
         $model = $this->findModel($id_review);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -114,6 +139,12 @@ class ReviewController extends Controller
      */
     public function actionDelete($id_review)
     {
+        if (!Yii::$app->user->can('deleteReview')) {
+            throw new \yii\web\ForbiddenHttpException(
+                'You do not have permission to delete reviews.'
+            );
+        }
+
         $this->findModel($id_review)->delete();
 
         return $this->redirect(['index']);
